Request based Proactive-Messaging and Channel/Member storage updated.
Channel/Member information is now stored and updated locally in separate arrays. Post requests to "/api/MessageRequest" and "/api/ChannelRequest" require a User/Channel name field which will be matched to the user/channel in the local array so the information for the new conversation can be pulled. The second field will be the message string to be sent.